server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9
pyghidra>=1.0.0
pydantic>=2.10.0
python-dotenv>=1.0.0

# Gemini API for code cleanup (new SDK)
google-genai>=1.0.0

# Google Cloud for Vertex AI integration
google-auth>=2.25.0
httpx>=0.27.0

# LLM4Decompile 1.3B dependencies
# Note: torch is installed separately in Dockerfile for CPU/GPU flexibility
transformers>=4.36.0
accelerate>=0.25.0

# OpenAI (commented out for now - will add back for GPT-4o integration later)
# openai==1.12.0
